+++ lib/spandx/ruby/parsers/gemfile_lock.rb
@@ -14,9 +14,9 @@ module Spandx
         def parse(lockfile)
           dependencies_from(lockfile).map do |specification|
             ::Spandx::Core::Dependency.new(
+              package_manager: :rubygems,
               name: specification.name,
               version: specification.version.to_s,
-              licenses: licenses_for(specification),
               meta: specification
             )
           end
@@ -31,16 +31,6 @@ module Spandx
               .new(content.sub(STRIP_BUNDLED_WITH, ''))
               .specs
           end
-        end
-
-        def licenses_for(specification)
-          rubygems
-            .licenses_for(specification.name, specification.version.to_s)
-            .map { |x| catalogue[x] }
-        end
-
-        def rubygems
-          @rubygems ||= Spandx::Ruby::Gateway.new
         end
       end
     end