+++ spec/unit/js/parsers/yarn_spec.rb
@@ -1,22 +1,15 @@
 # frozen_string_literal: true
 
 RSpec.describe Spandx::Js::Parsers::Yarn do
-  subject { described_class.new(catalogue: catalogue) }
-
-  let(:catalogue) { Spandx::Spdx::Catalogue.from_file(fixture_file('spdx/json/licenses.json')) }
+  subject { described_class.new }
 
   describe '#parse small lock file' do
     let(:lockfile) { fixture_file('js/yarn/short_yarn.lock') }
-    let(:result) do
-      VCR.use_cassette('js/yarn/short_yarn.lock') do
-        subject.parse(lockfile)
-      end
-    end
+    let(:result) { subject.parse(lockfile) }
 
     specify { expect(result.size).to eq(2) }
     specify { expect(result.first.name).to eq('babel') }
     specify { expect(result.first.version).to eq('6.23.0') }
-    specify { expect(result.first.licenses.map(&:id)).to match_array(['MIT']) }
   end
 
   describe '#invalid lock file' do
@@ -28,13 +21,8 @@ RSpec.describe Spandx::Js::Parsers::Yarn do
   describe '#parse long lock file' do
     let(:lockfile) { fixture_file('js/yarn/long_yarn.lock') }
     let(:expected_dependencies) { fixture_file_content('js/yarn/lol').lines.map(&:chomp) }
-    let(:result) do
-      VCR.use_cassette('js/yarn/long_yarn.lock') do
-        subject.parse(lockfile)
-      end
-    end
+    let(:result) { subject.parse(lockfile) }
 
     specify { expect(result.map { |x| "#{x.name}@#{x.version}" }) .to match_array(expected_dependencies) }
-    specify { expect(result.size).to eq(53) }
   end
 end