+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add maven index.
 - Add support for parsing `yarn.lock` files.
 - Add support for parsing `package-lock.json` files.
+- Add `--pull` option to fetch latest cache before scan.
 
 ### Changed
 - Change the default `--format` to `table` for the scan command.